+++ src/markehme/factionsplus/listeners/TeleportsListener.java
@@ -44,7 +44,7 @@ public class TeleportsListener implements Listener {
 		if ( !plugin.isEnabled() ) {
 			return;
 		}
-		if ( !Config.isInited() ) {
+		if ( !Config.isLoaded() ) {
 			throw FactionsPlusPlugin.bailOut( "bad call order while java coding, call this after config is loaded" );
 		}